Implemented adjoint output sensitivity, discovered interpolation problems
There doesn't seem to be a good way to control how CVODES performs state interpolation during the backward integration phase, which results in unavoidable interpolation over discontinuities. There has been a small amount of activity on the SUNDIALS mailing list about this, and it appears that one would need to make changes to CVodeReInit. This is more work than I had anticipated.